refactor: rename AuthTokenManagers::Custom to AuthTokenManager, drop .custom factory
`Neo4j::Driver::AuthTokenManager.new(get_token:, handle_security_exception:)`
is the public entry point on both impls. The old factory shortcut
`AuthTokenManagers.custom(...)` was redundant — `.new(...)` carries
the same kw-args directly. Clients that need richer behaviour subclass
(or write a duck-typed class from scratch); the protocol is unchanged.

On JRuby the shared Ruby class shadows the Java interface name in the
public namespace; the Java interface stays reachable internally as
`Java::OrgNeo4jDriver::AuthTokenManager`, which is what the adapter and
GraphDatabase dispatch already use. AuthTokenManagers keeps `.basic` /
`.bearer` on JRuby (delegating to Java's factory); the MRI surface for
those will land when MRI implements managed-auth.
