+++ lib/jruby/neo4j/driver/ext/auth_token_managers.rb
@@ -5,12 +5,12 @@ module Neo4j
     module Ext
       # Mirrors Java's `AuthTokenManagers.basic(Supplier<AuthToken>)` /
       # `bearer(Supplier<AuthTokenAndExpiration>)` factories with Ruby
-      # keyword-arg surfaces. The `custom` factory returns the shared
-      # duck-typed `Internal::AuthTokenManagers::Custom` (same class
-      # MRI uses); the boundary into Java's `AuthTokenManager`
-      # interface is crossed by `Internal::AuthTokenManagerAdapter`
-      # inside `GraphDatabase.driver`, so client code never names a
-      # Java type.
+      # keyword-arg surfaces. For the full-callback case (`get_token` +
+      # `handle_security_exception`) clients use
+      # `Neo4j::Driver::AuthTokenManager.new(...)` directly — that's
+      # the shared duck-typed entry point, and the boundary into Java's
+      # `AuthTokenManager` interface is crossed by
+      # `Internal::AuthTokenManagerAdapter` inside `GraphDatabase.driver`.
       module AuthTokenManagers
         def basic(supplier:)
           super(supplier)
@@ -18,10 +18,6 @@ module Neo4j
 
         def bearer(supplier:)
           super(supplier)
-        end
-
-        def custom(get_token:, handle_security_exception:)
-          Neo4j::Driver::Internal::AuthTokenManagers::Custom.new(get_token, handle_security_exception)
         end
       end
     end
