+++ testkit-backend/requests/new_auth_token_manager.rb
@@ -11,7 +11,7 @@ module TestkitBackend
 
       def to_object
         manager = nil
-        manager = Neo4j::Driver::AuthTokenManagers.custom(
+        manager = Neo4j::Driver::AuthTokenManager.new(
           get_token: -> { get_token(manager.object_id) },
           handle_security_exception: ->(token, exc) { handle_security_exception(manager.object_id, token, exc) }
         )
